fix: satisfy full test suite typing checks
Declare read-only worker semaphore fields for mypy, pass the workspace DuckDB path to scratchpad path resolution as a string, and make the resource lease retry delay return type explicit. Update the DB init recovery regression to expect the current Alembic head after the sibling-order migration.

+++ backend/app/data_access/coordinator.py
@@ -335,7 +335,7 @@ class ResourceLeaseCoordinator:
         return await operation()
 
     def _retry_delay(self, attempt: int) -> float:
-        return self._lock_retry_base_delay_seconds * (2**attempt)
+        return float(self._lock_retry_base_delay_seconds * (2**attempt))
 
     @staticmethod
     def _is_sqlite_locked_error(exc: BaseException) -> bool:
